Took UserInputHandler into different file.
